Don't start redis when we detect the platform is running JRuby

+++ spec/spec_helper.rb
@@ -23,25 +23,27 @@ REDIS_PID  = 'redis.pid'  # can't be absolute
 REDIS_DUMP = 'redis.rdb'  # can't be absolute
 REDIS_RUNDIR = File.dirname(__FILE__)
 
-describe 'redis-server' do
-  it "starting redis-server on #{REDIS_HOST}:#{REDIS_PORT}" do
-    fork_pid = fork do
-      system "cd #{REDIS_RUNDIR} && (echo port #{REDIS_PORT}; echo logfile /dev/null; echo daemonize yes; echo pidfile #{REDIS_PID}; echo dbfilename #{REDIS_DUMP}) | #{REDIS_BIN} -"
+if !(defined?(RUBY_ENGINE) && RUBY_ENGINE == "jruby")
+  describe 'redis-server' do
+    it "starting redis-server on #{REDIS_HOST}:#{REDIS_PORT}" do
+      fork_pid = fork do
+        system "cd #{REDIS_RUNDIR} && (echo port #{REDIS_PORT}; echo logfile /dev/null; echo daemonize yes; echo pidfile #{REDIS_PID}; echo dbfilename #{REDIS_DUMP}) | #{REDIS_BIN} -"
+      end
+      fork_pid.should > 0
+      sleep 2
     end
-    fork_pid.should > 0
-    sleep 2 
   end
-end
 
-at_exit do
-  pidfile = File.expand_path REDIS_PID,  REDIS_RUNDIR
-  rdbfile = File.expand_path REDIS_DUMP, REDIS_RUNDIR
-  pid = File.read(pidfile).to_i
-  puts "=> Killing #{REDIS_BIN} with pid #{pid}"
-  Process.kill "TERM", pid
-  Process.kill "KILL", pid
-  File.unlink pidfile
-  File.unlink rdbfile if File.exists? rdbfile
+  at_exit do
+    pidfile = File.expand_path REDIS_PID,  REDIS_RUNDIR
+    rdbfile = File.expand_path REDIS_DUMP, REDIS_RUNDIR
+    pid = File.read(pidfile).to_i
+    puts "=> Killing #{REDIS_BIN} with pid #{pid}"
+    Process.kill "TERM", pid
+    Process.kill "KILL", pid
+    File.unlink pidfile
+    File.unlink rdbfile if File.exists? rdbfile
+  end
 end
 
 def raises_exception(&block)
